+++ src/components/site/authForm.tsx
@@ -5,6 +5,7 @@ import { useTranslations } from "next-intl";
 import { useRouter } from "next-intl/client";
 import { signIn } from "next-auth/react";
 import toast from "react-hot-toast";
+import { MD5 } from "crypto-js";
 import { Loader2 } from "lucide-react";
 import { cn } from "@/lib";
 import {
@@ -14,6 +15,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
@@ -26,11 +28,17 @@ export default function AuthForm() {
   const tAuth = useTranslations("auth");
 
   const inputRef = React.useRef<any>(null);
+  const accountRef = React.useRef<any>(null);
+  const passwordRef = React.useRef<any>(null);
 
+  // account, password
+  const [activeTab, setActiveTab] = React.useState("email");
   const [email, setEmail] = React.useState<string>("");
+  const [password, setPassword] = React.useState<string>("");
   const [loadingGithub, setLoadingGithub] = React.useState(false);
   const [loadingGoogle, setLoadingGoogle] = React.useState(false);
   const [loadingEmail, setLoadingEmail] = React.useState(false);
+  const [loadingAccount, setLoadingAccount] = React.useState(false);
 
   const onGithubLogin = async () => {
     setLoadingGithub(true);
@@ -48,6 +56,11 @@ export default function AuthForm() {
   };
 
   const onLogin = async () => {
+    if (activeTab === "email") return onEmailLogin();
+    if (activeTab === "account") return onAccountLogin();
+  };
+
+  const onEmailLogin = async () => {
     if (loadingEmail) return;
 
     if (!email?.trim()) {
@@ -82,10 +95,58 @@ export default function AuthForm() {
     }
   };
 
+  const onAccountLogin = async () => {
+    if (loadingAccount) return;
+
+    if (!email?.trim()) {
+      toast.error(tGlobal("please-enter"), { id: "account" });
+      accountRef.current?.focus();
+      return;
+    }
+
+    const reg = /^[a-zA-Z0-9_-]+@[a-zA-Z0-9_-]+(\.[a-zA-Z0-9_-]+)+$/;
+
+    if (!reg.test(email)) {
+      toast.error(tAuth("input-valid-email"), { id: "valid_email" });
+      accountRef.current?.focus();
+      return;
+    }
+
+    if (!password?.trim()) {
+      toast.error(tGlobal("please-enter"), { id: "account" });
+      passwordRef.current?.focus();
+      return;
+    }
+
+    try {
+      setLoadingAccount(true);
+      const res: any = await signIn("credentials", {
+        callbackUrl: "/",
+        redirect: false,
+        username: email,
+        password: MD5(password).toString(),
+      });
+
+      if (res.error) {
+        setLoadingAccount(false);
+        toast.error(tAuth("auth-failed"));
+        return;
+      }
+
+      toast.success(tAuth("auth-success"));
+      router.push(res.url);
+    } catch (error) {
+      setLoadingAccount(false);
+      console.log(error, "login failed");
+      toast.error(tAuth("auth-failed"));
+    }
+  };
+
   return (
     <Card
       className={cn("w-96 max-w-[calc(100vw-2rem)] relative z-20", {
-        "pointer-events-none": loadingEmail || loadingGithub || loadingGoogle,
+        "pointer-events-none":
+          loadingEmail || loadingAccount || loadingGithub || loadingGoogle,
       })}
     >
       <CardHeader>
@@ -96,12 +157,12 @@ export default function AuthForm() {
         {checkGithub() && (
           <Button
             variant="outline"
-            className="w-full justify-start gap-4 mb-2"
+            className="mb-2 w-full gap-4 justify-start"
             disabled={loadingGithub}
             onClick={onGithubLogin}
           >
             {loadingGithub ? (
-              <Loader2 className="h-4 w-4 animate-spin" />
+              <Loader2 className="h-4 animate-spin w-4" />
             ) : (
               <Icon
                 className="text-black dark:text-white"
@@ -115,12 +176,12 @@ export default function AuthForm() {
         {checkGoogle() && (
           <Button
             variant="outline"
-            className="w-full justify-start gap-4"
+            className="w-full gap-4 justify-start"
             disabled={loadingGoogle}
             onClick={onGoogleLogin}
           >
             {loadingGoogle ? (
-              <Loader2 className="h-4 w-4 animate-spin" />
+              <Loader2 className="h-4 animate-spin w-4" />
             ) : (
               <Icon icon="google_line" size={20} />
             )}
@@ -128,31 +189,66 @@ export default function AuthForm() {
           </Button>
         )}
         {checkEmail() && (checkGithub() || checkGoogle()) && (
-          <div className="flex items-center my-8">
-            <div className="h-[1px] flex-1 bg-[rgba(0,0,0,0.16)]"></div>
+          <div className="flex my-8 items-center">
+            <div className="bg-[rgba(0,0,0,0.16)] flex-1 h-[1px]"></div>
             <div className="mx-4 text-[13px]">{tGlobal("or")}</div>
-            <div className="h-[1px] flex-1 bg-[rgba(0,0,0,0.16)]"></div>
+            <div className="bg-[rgba(0,0,0,0.16)] flex-1 h-[1px]"></div>
           </div>
         )}
         {checkEmail() && (
           <>
-            <div className="flex flex-col space-y-1.5 mb-4">
-              <Label htmlFor="name">{tAuth("email-address")}</Label>
-              <Input
-                ref={inputRef}
-                id="name"
-                className="h-[38px]"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-            </div>
+            <Tabs value={activeTab} onValueChange={setActiveTab}>
+              <TabsList className="w-full grid grid-cols-2">
+                <TabsTrigger value="email">{tAuth("email-login")}</TabsTrigger>
+                <TabsTrigger value="account">
+                  {tAuth("account-login")}
+                </TabsTrigger>
+              </TabsList>
+              <TabsContent value="email">
+                <div className="flex flex-col space-y-1.5 my-4">
+                  <Label htmlFor="email">{tAuth("email-address")}</Label>
+                  <Input
+                    ref={inputRef}
+                    id="email"
+                    className="h-[38px]"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
+                  />
+                </div>
+              </TabsContent>
+              <TabsContent value="account">
+                <div className="my-4 w-full grid gap-4 items-center">
+                  <div className="flex flex-col space-y-1.5">
+                    <Label htmlFor="account">{tAuth("account")}</Label>
+                    <Input
+                      ref={accountRef}
+                      className="h-[38px]"
+                      id="account"
+                      value={email}
+                      onChange={(e) => setEmail(e.target.value)}
+                    />
+                  </div>
+                  <div className="flex flex-col space-y-1.5">
+                    <Label htmlFor="password">{tAuth("password")}</Label>
+                    <Input
+                      ref={passwordRef}
+                      className="h-[38px]"
+                      id="password"
+                      type="password"
+                      value={password}
+                      onChange={(e) => setPassword(e.target.value)}
+                    />
+                  </div>
+                </div>
+              </TabsContent>
+            </Tabs>
             <Button
               className="w-full"
               onClick={onLogin}
-              disabled={loadingEmail}
+              disabled={loadingEmail || loadingAccount}
             >
-              {loadingEmail && (
-                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+              {(loadingEmail || loadingAccount) && (
+                <Loader2 className="h-4 mr-2 animate-spin w-4" />
               )}
               {tGlobal("continue")}
             </Button>