+++ CHANGE_LOG.md
@@ -2,13 +2,17 @@
 
 ## v0.9.1
 
-> 2023-09-04
+> 2023-09-05
 
 ### Fixed
 
 - Fixed the issue of incorrect authentication logic after configuring a custom API Key when deploying a project
 - Fixed the issue of garbled characters in some cases when replying
 - Fixed [#8](https://github.com/LTopx/Le-AI/issues/8)
+
+### Add
+
+- Support for account/password login
 
 ## v0.9.0